Modificato idRuolo con valore int

+++ src/main/java/it/nesea/albergo/utente_ruolo/dto/request/CreaUtenteDto.java
@@ -26,6 +26,6 @@ public class CreaUtenteDto implements Serializable {
     private String cognome;
 
     @Min(1)
-    private byte idRuolo;
+    private int idRuolo;
 
 }
